+++ maskrcnn_benchmark/modeling/roi_heads/mask_head/inference.py
@@ -76,7 +76,7 @@ class MaskPostProcessorCOCOFormat(MaskPostProcessor):
         for result in results:
             masks = result.get_field("mask").cpu()
             rles = [
-                mask_util.encode(np.array(mask[0, :, :, np.newaxis], order="F"))[0]
+                mask_util.encode(np.array(mask[0, :, :, np.newaxis], dtype=np.uint8, order="F"))[0]
                 for mask in masks
             ]
             for rle in rles:
